requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
pydantic-settings
python-multipart
fastapi-mail
slowapi

# testing
pytest
pytest-asyncio
httpx
